Output facilities, console messages, and fail-fast protections
Implemented new commandline output options.  Now there is a clean separation between console messages for the user, various output files, a logging facility (useful for both users and for logging runtime errors that can be sent to us for debugging), and a debugging option that I will remove once we go fully public.  In addition, code was added to allow the program to fail fast and meaningfully in a number of scenarios, e.g. malformed input files.