Precomputed ancestral reconstruction feature
Added -u and --use-precomputed-anc-recon to the command line.  This new feature allows the user to utilize a precomputed ancestral reconstruction fasta and newick file.  When the option is turned on, treetime is bypassed completely and the program understands that it is now looking at a fasta file with ancestral genotypes as well as a newick with internal nodes labelled.  In addition, I've updated the program to output the newick tree that I parse out from treetime's nexus output file.  This allows users to easily reuse this ancestral_tree.newick file (along with its accompanying ancestral_sequences.fasta file) for subsequent poutine sessions to bypass ancestral reconstruction and more quickly explore other GWAS settings and/or phenotypes.