# alleles at each site
Program is now identifying the # alleles at each segregating site, and also reporting this metric in the log file (and debug file).  This is particularly useful as a sanity check to see which sites are being assayed by poutine (only biallelic sites for now).  I anticipate users not considering this point, and thus seeing the # non-biallelic sites can potentially alert the user to either problems in their dataset and/or # sites that are not being considered by poutine.

This need came from analyzing the discovery set where a substantial number of sites turned out to be monomorphic (due likely to subclonal heterogeneity being picked up by the variant caller).  A future version may include the capability of considering multiallelic sites perhaps using a multinomial test.