Minor code cleanup
It's been a long few days of coding and thinking about the algo.  Few hours of sleep left the code needing a little clean up, so here it is.  There are still items I'm aware of that I'd like to cleanup, but the focus now is on the quality of the algo and also making it easy for users.  All changes per usual have been tested before commits.  More coffee now . . .

- removed TODO (GENO MISSINGNESS) comments
- commented out a few lines that are not needed
- fixed a few log formatting issues